bdp_assignment/pub_sub_broker/requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.5.2
protobuf==4.23.1
tqdm==4.65.0
